Fixes to OS lookup
